requirements.txt
django==1.8.1
django-extensions==1.5.3
django-debug-toolbar==1.3.0
django-gravatar2==1.3.0
django-cors-headers==1.0.0
django-tag-parser==2.1
django-cacheops==2.3.2
django-redis==4.0.0
dj-database-url==0.3.0

djangorestframework==3.1.1
djangorestframework-camel-case==0.2.0
django-rest-swagger==0.3.3

psycopg2==2.6
pip-tools==1.1.2
pyyaml==3.11
pillow==2.8.1
sorl-thumbnail==12.2
gunicorn==19.3.0
